requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
arrow==1.3.0
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
blessed==1.20.0
boto3==1.28.62
botocore==1.31.62
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
contourpy==1.1.1
croniter==1.4.1
cycler==0.12.1
dateutils==0.6.12
deepdiff==6.6.0
docker-pycreds==0.4.0
exceptiongroup==1.1.3
fastapi==0.103.2
filelock==3.9.0
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.4.0
gitdb==4.0.10
GitPython==3.1.37
h11==0.14.0
h5py==3.10.0
idna==3.4
inquirer==3.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
kiwisolver==1.4.5
lightning==2.0.9.post0
lightning-cloud==0.5.41
lightning-utilities==0.9.0
markdown-it-py==3.0.0
MarkupSafe==2.1.2
mat73==0.62
matplotlib==3.8.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.0
numpy==1.24.1
opencv-python-headless==4.8.1.78
ordered-set==4.1.0
packaging==23.2
pathtools==0.1.2
Pillow==9.3.0
protobuf==4.24.4
psutil==5.9.5
pydantic==2.1.1
pydantic_core==2.4.0
Pygments==2.16.1
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-editor==1.0.4
python-multipart==0.0.6
pytorch-lightning==2.0.9.post0
pytz==2023.3.post1
PyYAML==6.0.1
readchar==4.0.5
requests==2.28.1
rich==13.6.0
s3transfer==0.7.0
scikit-learn==1.3.1
scipy==1.11.3
sentry-sdk==1.32.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
starlette==0.27.0
starsessions==1.3.0
sympy==1.12
threadpoolctl==3.2.0
torch==2.1.0+cu118
torchmetrics==1.2.0
torchvision==0.16.0+cu118
tqdm==4.66.1
traitlets==5.11.2
triton==2.1.0
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
urllib3==1.26.13
uvicorn==0.23.2
wandb==0.15.12
wcwidth==0.2.8  
websocket-client==1.6.4
websockets==11.0.3
yarl==1.9.2
--find-links https://download.pytorch.org/whl/cu118/torch_stable.html